+++ test/index.cjs
@@ -1,0 +1,10 @@
+const path = require('upath');
+const { runCommand } = require('./utils.cjs');
+
+const cwd = path.resolve(__dirname, '..');
+process.cwd = () => cwd;
+
+runCommand('node', ['test/hexo-starter.cjs'], {
+  stdio: 'inherit',
+  cwd
+});