+++ test/sitemap/yoastseo.ts
@@ -1,5 +1,7 @@
-import { bindProcessExit, fs, path } from 'sbg-utility';
-import { baseSite, envHexo } from '../env';
+import fs from 'fs';
+import { bindProcessExit } from 'sbg-utility';
+import path from 'upath';
+import { baseSite, envHexo } from '../env.cjs';
 
 // test only activate yoastseo sitemap